requirements.txt
Django==3.1.7

django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djoser==2.2.3
social-auth-app-django==5.4.2

django-ckeditor==6.7.1

pytz==2024.1

Pillow==10.4.0
mutagen==1.47.0 

psycopg2==2.9.2  

argon2-cffi==23.1.0
django-environ==0.11.2

stripe==10.8.0
braintree==4.29.0

pyJWT==2.9.0

boto3==2.49.0
django-storages==1.14.4

markdown==3.7
gunicorn==23.0.0
whitenoise==6.7.0

requests==2.32.3